% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakagami.R
\name{Nakagami}
\alias{Nakagami}
\alias{dnaka}
\alias{pnaka}
\alias{qnaka}
\alias{rnaka}
\title{The Nakagami Density}
\usage{
dnaka(x, shape, scale, log = FALSE)

pnaka(q, shape, scale, lower.tail = TRUE, log.p = FALSE)

qnaka(p, shape, scale, lower.tail = TRUE, log.p = FALSE)

rnaka(n, shape, scale)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{shape}{vector of shape parameters greater than `1/2`.}

\item{scale}{vector of positive scale parameters.}

\item{log, log.p}{logical; if `TRUE`, probabilities `p` are given as `log(p)`.}

\item{lower.tail}{logical; if `TRUE` (default), probabilities are
\eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`, the length is taken to
be the number required.}
}
\value{
`dnaka` gives the density, `pnaka` gives the distribution function,
   `qnaka` gives the quantile function and `rnaka` generates random deviates.

   The length of the result is determined by `n` for `rnaka`, and is the
   maximum of the lengths of the numerical arguments for the other functions.

   The numerical arguments other than `n` are recycled to the length of the
   result.
}
\description{
Density, distribution function, quantile function and random generation for
   the Nakagami distribution with parameters `shape` and `scale`.
}
\details{
The Nakagami distribution with parameters `shape = m` and `scale = Omega` has
   density \deqn{2m^m/{\Gamma(m)\Omega^m} x^(2m-1)e^(-m/\Omega x^2)} for
   \eqn{x \geq 0}, \eqn{m \geq 1/2} and \eqn{Omega > 0}.

If \eqn{Y} is [Gamma][stats::GammaDist] distributed with `shape = m` and
   `rate = m/Omega` then \eqn{X = \sqrt Y} is Nakagami distributed
   with `shape = m` and `scale = Omega`.
}
\references{
Nakagami, N. 1960. "The M-Distribution, a General Formula of
   Intensity of Rapid Fading." In Statistical Methods in Radio Wave
   Propagation: Proceedings of a Symposium Held at the University of
   California, edited by William C. Hoffman, 3-36. Permagon Press.
}
