% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{suppress_olw}
\alias{suppress_olw}
\title{Suppress object length incompatibility warnings}
\usage{
suppress_olw(expr)
}
\arguments{
\item{expr}{expression to be evaluated.}
}
\description{
Suppress object length incompatibility warnings
}
